+++ setup.py
@@ -22,7 +22,7 @@ setup(name='nucypher',
       include_package_data=True,
       entry_points='''
                    [console_scripts]
-                   nucypher-cli=nucypher_cli.main:cli
-                   run-ursula=nucypher_cli.run_ursula:run_ursula
+                   nucypher-cli=cli.main:cli
+                   run-ursula=cli.run_ursula:run_ursula
                    '''
       )
