+++ tests/network/test_network_upgrade.py
@@ -6,7 +6,6 @@ from cryptography.hazmat.primitives import serialization
 from twisted.internet import threads
 
 from nucypher.characters import Ursula
-from nucypher.utilities.blockchain import make_ursulas
 from nucypher.utilities.sandbox import make_ursulas
